tests/submission/W3C/NavigationTiming/monolitic.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Monolitic Clock</title>
<link rel="author" title="W3C" href="http://www.w3.org/" />
<link rel="help" href="http://www.w3.org/TR/navigation-timing/#mono-clock" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<h1>Description</h1>
<p>The difference between any two chronologically recorded timing attributes must never be negative.</p>

<script>
function assert_greater_than(property1, property2) {
    test(function() {
	    var timing = window.performance.timing;
            assert_true( (timing[property1] >= timing[property2]) );
	}, property1 + ">=" + property2 );
}

setup({ explicit_done: true });

// "secureConnectionStart",

function nav_timing() {
    assert_greater_than("unloadEventEnd", "unloadEventStart");
    assert_greater_than("redirectEnd", "redirectStart");
    assert_greater_than("fetchStart", "navigationStart");
    assert_greater_than("domainLookupStart", "fetchStart");
    assert_greater_than("domainLookupEnd", "domainLookupStart");
    assert_greater_than("connectStart", "domainLookupEnd");
    assert_greater_than("connectEnd", "connectStart");
    assert_greater_than("requestStart", "connectEnd");
    assert_greater_than("responseStart", "requestStart");
    assert_greater_than("responseEnd", "responseStart");
    assert_greater_than("domLoading", "responseEnd");
    assert_greater_than("domInteractive", "domLoading");
    assert_greater_than("domContentLoadedEventStart", "domInteractive");
    assert_greater_than("domContentLoadedEventEnd", "domContentLoadedEventStart");
    assert_greater_than("domComplete", "domContentLoadedEventEnd");
    assert_greater_than("loadEventStart", "domComplete");
    assert_greater_than("loadEventEnd", "loadEventStart");
    done();
}
function onload() {
  setTimeout(nav_timing, 100);
}

on_event(window, "load", onload);

</script>

<div id="log"></div>

</body>
</html>
